Restructure gdpr suff in project and remove privacy route.
