+++ src/components/gdpr/GDPR.vue
