+++ src/components/gdpr/Privacy.vue
