+++ src/components/gdpr/PrivacyDialog.vue
